refactor imports of components
