+++ src/components/index.js
@@ -1,0 +1,15 @@
+import LanguageSwitcher from './language-switcher';
+import Logo from './logo';
+import Menu from './menu';
+import Spinner from './spinner';
+import Table from './table';
+import Tree from './tree';
+
+export {
+  LanguageSwitcher,
+  Logo,
+  Menu,
+  Spinner,
+  Table,
+  Tree
+}
