+++ src/components/tree/tree-node.js
@@ -13,7 +13,7 @@ function TreeNode({ data, level = 1 }) {
             <div
               className={
                 classNames(
-                  'tree__node', 'hidden',
+                  'tree__node hidden',
                   { 'tree__node_no-children': !item?.children?.length }
                 )
               }
