+++ src/pages/table-page/table-page.js
@@ -2,8 +2,7 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { fetchFamilyTable } from '../../store/reducer';
 import { useTranslation } from 'react-i18next';
-import Spinner from '../../components/spinner';
-import Table from '../../components/table';
+import { Spinner, Table } from '../../components';
 import { configTable } from '../../config';
 import './table-page.scss';
